FIRE-236 rename id to nickname
모든 ID라고 표시된 부분을 nickname으로 수정
